Population Growth Application with initial number of organisms, birthrate, deathrate and randomized catastrophies
Application that predicts the approximate size of a population of organisms allowing the user to enter the starting number of organisms, the average daily population increase (as a percentage), and the number of days the organisms will be left to multiply.

Used a ListBox to display the population size for each day of the simulation.

If the user clicks the Save button, save the current prediction to a file.